Add files via upload
As a starting point, I chose the flat panel photobioreactor design for this project due to its simplicity and efficient light exposure. The design also is easy to understand and sketch on Fusion 360 in the upcoming weeks. However this is a preliminary choice. As I continue researching on other models (eg: tubular, vertical column), I will evaluate whether a different design could have a better performance or feasible for my specific project.